altos586: add Altos 586 emulation
This is a draft intended to get early feedback before I spend too much
effort getting XENIX to run on this machine. This was a C++ learning
exercise for me and I'm pretty sure some of the code stinks. I might have
cargo culted a plenty of stuff without grasping an overall idea.

Aside from language issues, the computer is a bit of an oddball (perhaps
many of the machines emulated are) and I'm not sure if I've modeled it
right.

The system has main board with a 8086 CPU, a Z80 peripheral processor,
a MMU, 512K RAM (expansion slot for another 512K which I don't have) and
one quite Multibus-y expansion port. Into the expansion port connects a
hard drive controller based around a 8089 in remote configuration.

The MMU policies access to all the memory and I/O peripherals and its
emulation has been giving me headaches. Here's a short summary of what
it does:

1.) Memory is split into 256 pages of 4096 bytes with mapping and
    protection bits that apply to everyone accessing the memory
    (be it CPU or other bus masters).
2.) There's no mapping or protection bits for I/O.
3.) Accesses from CPU are treated differently than those from the
    other bus masters (Z80 IOP or 8089 on HDC board)
4.) MMU can be told the CPU is running privileged ("system mode"):
    * Write protection still applies
    * All reads from CPU are allowed
    * All I/O access from CPU is allowed
5.) MMU can be told the CPU is running unprivileged ("user mode"):
    * Read and Write access violations from CPU cause a NMI
    * Disabling interrupts (S5 going low I guess) cause a NMI
    * I/O attempts from CPU cause a regular interrupt IRQ0
    * All of the above switch to system mode
6.) Pages can be configured to disallow access from other bus
    masters. Violations cause a NMI.
7.) Other bus masters can access I/O space beyond 0400h (no mapping
    and protection bits for I/O) unhindered, but not anything below.

The peculiar thing is that the MMU needs to tell whether the access is
coming from the CPU in user mode, CPU in system mode or an external bus
master.

I've solved this by splitting the MMU into a device instance separate
from the machine object and have them both provide
device_memory_interface -- the machine object provides unobstructed
access for MMU's needs, while the MMU provides mapped access for other
bus masters. There are separate routines to deal with accesses from the
main 8086 CPU along with the MMU keeping track of CPU's mode.
Could there be a better way?

Note the MMU is similar to that of 8600. The format of the mapping
tables is a little different. I've not reused the code because it
doesn't seem to deal with the problem of policing access from external
devices.

Apart from that the code is littered with TODOs. They generally serve
as reminders for whoever is reading that I couldn't do any better at
the time of writing. All feedback is appreciated, but I'm especially
grateful for directions at those points.