cpu/i86: make "out dx,al" output masked AX on data bus
The diagnostic software for the Altos 586 computer uses the
"out dx,al" (0EEh) opcode instead of "out dx,ax" (0EFh) when writing 16-bit
words to its MMU's memory mapping table.

It works just fine, presumably because 1.) the MMU doesn't observe the BHE
signal and 2.) the processor actually places the upper AH part of the AX
register on the bus when accessing an even address.

The commit 354656e55b84 ('i186: Peripheral register refinements')
suggests that this might be consistent with what seems to be a well-known
behavior of an 80186 core accessing the internal peripherals.